+++ project/app-domain/src/main/java/mobi/lab/sample/domain/di/UseCaseModule.kt
@@ -1,9 +1,0 @@
-package mobi.lab.sample.domain.di
-
-import dagger.Module
-
-/**
- * UseCases are provided by constructor injection from usecases package
- */
-@Module
-object UseCaseModule